Move to new attributes folder
